zram: switch to 64-bit hash for dedup The original dedup code does not handle collision from the observation that it practically does not happen.
For additional peace of mind, use a bigger hash size for reducing the
possibility of collision even further.

+++ drivers/block/zram/zram_dedup.h
@@ -8,9 +8,9 @@ u64 zram_dedup_dup_size(struct zram *zram);
 u64 zram_dedup_meta_size(struct zram *zram);
 
 void zram_dedup_insert(struct zram *zram, struct zram_entry *new,
-				u32 checksum);
+				u64 checksum);
 struct zram_entry *zram_dedup_find(struct zram *zram, struct page *page,
-				u32 *checksum);
+				u64 *checksum);
 
 void zram_dedup_init_entry(struct zram *zram, struct zram_entry *entry,
 				unsigned long handle, unsigned int len);
